adds build system for css & js, coordianted via npm scripts
