--- public/style.css
@@ -1,51 +1,0 @@
-h1, h2, h3, h4, h5, h6 {
-  font-family: alfa slab one, helvetica, arial;
-}
-
-p, div {
-  font-family: lucida grande, arial;
-}
-
-.headline {
-  font-family: alfa slab one, helvetica, arial;
-}
-
-.sans-serif {
-  font-family: lucida grande, arial;
-}
-
-.noselect {
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-
-.mw5 {
-  max-width: 20rem;
-}
-
-.bg-navy {
-  background-color: #3C4366;
-}
-
-.navy {
-  color: #3C4366;
-}
-
-.b--navy {
-  border-color: #3C4366;
-}
-
-.bg-red {
-  background-color: #E76464;
-}
-
-.red {
-  color: #E76464;
-}
-
-.b--red {
-  border-color: #E76464;
-}
