+++ src/css/index.css
@@ -1,0 +1,51 @@
+/* import tachyons modules */
+@import 'tachyons-custom/src/_box-sizing';
+@import 'tachyons-custom/src/_images';
+@import 'tachyons-custom/src/_background-size';
+@import 'tachyons-custom/src/_background-position';
+@import 'tachyons-custom/src/_outlines';
+@import 'tachyons-custom/src/_borders';
+@import 'tachyons-custom/src/_border-colors';
+@import 'tachyons-custom/src/_border-radius';
+@import 'tachyons-custom/src/_border-style';
+@import 'tachyons-custom/src/_border-widths';
+@import 'tachyons-custom/src/_box-shadow';
+@import 'tachyons-custom/src/_code';
+@import 'tachyons-custom/src/_coordinates';
+@import 'tachyons-custom/src/_clears';
+@import 'tachyons-custom/src/_display';
+@import 'tachyons-custom/src/_flexbox';
+@import 'tachyons-custom/src/_floats';
+@import 'tachyons-custom/src/_font-family';
+@import 'tachyons-custom/src/_font-style';
+@import 'tachyons-custom/src/_font-weight';
+@import 'tachyons-custom/src/_forms';
+@import 'tachyons-custom/src/_heights';
+@import 'tachyons-custom/src/_letter-spacing';
+@import 'tachyons-custom/src/_line-height';
+@import 'tachyons-custom/src/_links';
+@import 'tachyons-custom/src/_lists';
+@import 'tachyons-custom/src/_max-widths';
+@import 'tachyons-custom/src/_widths';
+@import 'tachyons-custom/src/_overflow';
+@import 'tachyons-custom/src/_position';
+@import 'tachyons-custom/src/_opacity';
+@import 'tachyons-custom/src/_skins';
+@import 'tachyons-custom/src/_skins-pseudo';
+@import 'tachyons-custom/src/_spacing';
+@import 'tachyons-custom/src/_tables';
+@import 'tachyons-custom/src/_text-decoration';
+@import 'tachyons-custom/src/_text-align';
+@import 'tachyons-custom/src/_text-transform';
+@import 'tachyons-custom/src/_type-scale';
+@import 'tachyons-custom/src/_typography';
+@import 'tachyons-custom/src/_utilities';
+@import 'tachyons-custom/src/_visibility';
+@import 'tachyons-custom/src/_white-space';
+@import 'tachyons-custom/src/_vertical-align';
+@import 'tachyons-custom/src/_hovers';
+@import 'tachyons-custom/src/_z-index';
+
+/* custom variables */
+@import './_variables';
+@import './_custom';
